Revert using shlex for BNCS command processing
It was causing a lot of problems with quotes in normal messages with little actual benefit. For whispering users with spaces in their name (which is afaik still not supported on the server) you can use '?' in place of the space.

Also adds an error catch on the `/capi send` command.